requirement.txt
psycopg2
django
djangorestframework
south
django-extensions
django-rest-swagger